Separate out ui handler under build tag
